requirements.txt
streamlit
streamlit-cropper
pillow
torch
transformers
huggingface-hub
python-dotenv
python-doctr
opencv-python-headless
python-doctr[torch]